Create JamfPro-AdHocAppConfigDeploy.sh
This is a very specific script. It can deploy a managed App Configuration for a installed application for a specific set of devices (sorted into groups). The script can send app configurations that are either seperate or the same to the app config set on the App record in Jamf Pro